moved readstats into scripts/
