+++ doc/user/scripts.txt
@@ -126,6 +126,9 @@ Read handling: interleaving, splitting, etc.
 .. autoprogram:: interleave-reads:get_parser()
         :prog: interleave-reads.py
 
+.. autoprogram:: readstats:get_parser()
+        :prog: readstats.py
+
 .. autoprogram:: sample-reads-randomly:get_parser()
         :prog: sample-reads-randomly.py
